Add a fun little emoji filter as a bonus feature.
I went with this as a bonus feature because I think it was such a strong idea Slack came up with
back when I first used slack. No need for ActionText wysiwygs if you can just parse the colon
delimiters correctly. Also it was going to be simple and not too hard.